mms_be/requirements.txt
asgiref==3.9.2
cfgv==3.4.0
diff-match-patch==20241021
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
django-import-export==4.3.10
djangorestframework==3.16.1
filelock==3.19.1
identify==2.6.14
nodeenv==1.9.1
platformdirs==4.4.0
pre_commit==4.3.0
psycopg==3.2.10
psycopg2==2.9.10
python-dotenv==1.1.1
PyYAML==6.0.3
ruff==0.13.2
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
virtualenv==20.34.0
